extract AppMode and AppState
